routes to leisure updated

+++ main.py
@@ -139,40 +139,53 @@ def show_leisure():
 
 @app.route('/leisure', methods=['GET'])
 def show_leisure():
-    l_type = 'MUSEUM'
-    leisure = Leisure(l_type).get_by_id(710)
-    ratings = Rating.get_ratings_by_leisure('leisure', 710)
-    return render_template('leisure.html', l=leisure, ratings=ratings)
-
-
-@app.route('/saverating', methods=['GET', 'POST'])
-def save_rating():
-    if request.method == 'POST':
-        r_Id = request.form['id']
-        grade = int(request.form['grade'])
-        description = request.form['description']
-        l_Id = int(request.form['leisure'])
-        l_type = request.form['type']
-        user = session['current_user_uid']
-        if r_Id == 'null':
-            Rating.create_rating(grade, description, l_Id, user)
-        else:
-            data = {"grade": grade, "description": description}
-            Rating.update_rating(r_Id, data)
+    if 'current_user_uid' not in session:
+        return redirect(url_for('start'))
+    else:
+        #l_type = 'MUSEUM'
+        l_type = request.args.get('type')
+        l_Id = int(request.args.get('id'))
         leisure = Leisure(l_type).get_by_id(l_Id)
         ratings = Rating.get_ratings_by_leisure('leisure', l_Id)
         return render_template('leisure.html', l=leisure, ratings=ratings)
 
 
+@app.route('/saverating', methods=['GET', 'POST'])
+def save_rating():
+    if 'current_user_uid' not in session:
+        return redirect(url_for('start'))
+    else:
+        if request.method == 'POST':
+            r_Id = request.form['id']
+            grade = int(request.form['grade'])
+            description = request.form['description']
+            l_Id = int(request.form['leisure'])
+            l_type = request.form['type']
+            user = session['current_user_uid']
+            if r_Id == 'null':
+                Rating.create_rating(grade, description, l_Id, user)
+            else:
+                data = {"grade": grade, "description": description}
+                Rating.update_rating(r_Id, data)
+            #leisure = Leisure(l_type).get_by_id(l_Id)
+            #ratings = Rating.get_ratings_by_leisure('leisure', l_Id)
+            #return render_template('leisure.html', l=leisure, ratings=ratings)
+            return redirect(url_for('show_leisure', id = l_Id, type = l_type))
+
+
 @app.route('/deleterating', methods=['GET', 'POST'])
 def delete_rating():
-    r_Id = request.form['id']
-    l_Id = int(request.form['leisure'])
-    l_type = request.form['type']
-    Rating.delete_rating(r_Id)
-    leisure = Leisure(l_type).get_by_id(l_Id)
-    ratings = Rating.get_ratings_by_leisure('leisure', l_Id)
-    return render_template('leisure.html', l=leisure, ratings=ratings)
+    if 'current_user_uid' not in session:
+        return redirect(url_for('start'))
+    else:
+        r_Id = request.form['id']
+        l_Id = int(request.form['leisure'])
+        l_type = request.form['type']
+        Rating.delete_rating(r_Id)
+        #leisure = Leisure(l_type).get_by_id(l_Id)
+        #ratings = Rating.get_ratings_by_leisure('leisure', l_Id)
+        #return render_template('leisure.html', l=leisure, ratings=ratings)
+        return redirect(url_for('show_leisure', id = l_Id, type = l_type))
 
 '''
 @app.route('/leisure/create', methods=['GET', 'POST'])
